+++ src/eslint-plugin-adeira/src/rules/__tests__/valid-test-folder.test.js
@@ -5,7 +5,7 @@ const RuleTester = require('eslint').RuleTester;
 const rule = require('../valid-test-folder');
 
 const ruleTester = new RuleTester({
-  parser: require.resolve('@babel/eslint-parser'),
+  parser: require.resolve('hermes-eslint'),
 });
 
 ruleTester.run('valid-test-folder', rule, {